+++ apps/web/src/views/Swap/MMLinkPools/components/MMCommitButton.tsx
@@ -1,5 +1,5 @@
 import { useTranslation } from '@pancakeswap/localization'
-import { Currency, CurrencyAmount, TradeType } from '@pancakeswap/sdk'
+import { Currency, CurrencyAmount, Token, TradeType } from '@pancakeswap/sdk'
 import { SmartRouterTrade } from '@pancakeswap/smart-router/evm'
 import { Button, useModal } from '@pancakeswap/uikit'
 import { useExpertMode } from '@pancakeswap/utils/user'
@@ -7,20 +7,16 @@ import { CommitButton } from 'components/CommitButton'
 import ConnectWalletButton from 'components/ConnectWalletButton'
 import SettingsModal, { withCustomOnDismiss } from 'components/Menu/GlobalSettings/SettingsModal'
 import { SettingsMode } from 'components/Menu/GlobalSettings/types'
-import { usePermitOrApprove } from 'hooks/usePermitStatus'
 import { WrapType } from 'hooks/useWrapCallback'
 import { useCallback, useEffect, useMemo, useState } from 'react'
 import { Field } from 'state/swap/actions'
 import { logGTMClickSwapEvent } from 'utils/customGTMEventTracking'
-import { isUserRejected } from 'utils/sentry'
 import { Address } from 'viem'
 import { parseMMError } from 'views/Swap/MMLinkPools/utils/exchange'
 import { ConfirmSwapModal } from 'views/Swap/V3Swap/containers/ConfirmSwapModal'
-import { useConfirmModalState } from 'views/Swap/V3Swap/hooks/useConfirmModalState'
-import { TransactionRejectedError } from 'views/Swap/V3Swap/hooks/useSendSwapTransaction'
 import { MMTradeInfo } from '../hooks'
+import { useMMConfirmModalState } from '../hooks/useMMConfirmModalState'
 import { useSwapCallArgumentsV2 } from '../hooks/useSwapCallArguments'
-import { useSwapCallback } from '../hooks/useSwapCallback'
 import { MMRfqTrade } from '../types'
 
 const SettingsModalWithCustomDismiss = withCustomOnDismiss(SettingsModal)
@@ -72,63 +68,36 @@ SwapCommitButtonPropsType) {
   const [isExpertMode] = useExpertMode()
 
   const { t } = useTranslation()
-  // the callback to execute the swap
+  const [tradeToConfirm, setTradeToConfirm] = useState<SmartRouterTrade<TradeType> | undefined>(undefined)
+  const amountToApprove = useMemo(() => {
+    return mmTradeInfo?.slippageAdjustedAmounts[Field.INPUT]
+  }, [mmTradeInfo])
+  const mmSpender = useMemo(() => {
+    return mmTradeInfo?.routerAddress as Address | undefined
+  }, [mmTradeInfo])
   const swapCalls = useSwapCallArgumentsV2(rfqTrade.trade, rfqTrade.rfq ?? undefined, recipient ?? undefined)
-
-  const { callback: swapCallback, error: swapCallbackError } = useSwapCallback(rfqTrade.trade, recipient, swapCalls)
-  const [{ tradeToConfirm, swapErrorMessage, attemptingTxn, txHash }, setSwapState] = useState<{
-    tradeToConfirm?: SmartRouterTrade<TradeType>
-    attemptingTxn: boolean
-    swapErrorMessage: string | undefined
-    txHash: string | undefined
-  }>({
-    tradeToConfirm: undefined,
-    attemptingTxn: false,
-    swapErrorMessage: undefined,
-    txHash: undefined,
-  })
+  const { resetState, txHash, confirmState, confirmSteps, callToAction, errorMessage } = useMMConfirmModalState(
+    rfqTrade.trade,
+    swapCalls,
+    (recipient as Address) ?? null,
+    amountToApprove?.currency.isToken ? (amountToApprove as CurrencyAmount<Token>) : undefined,
+    mmSpender,
+  )
 
   // Handlers
 
-  const onConfirmSwap = useCallback(async () => {
-    // @TODO should mm trade add price impact limit?
-
-    if (!swapCallback) return undefined
-
-    try {
-      setSwapState({ attemptingTxn: true, tradeToConfirm, swapErrorMessage: undefined, txHash: undefined })
-      const result = await swapCallback()
-      setSwapState({ attemptingTxn: false, tradeToConfirm, swapErrorMessage: undefined, txHash: result.hash })
-      return result
-    } catch (error: any) {
-      console.error(error)
-      const userRejected =
-        (typeof error !== 'string' && isUserRejected(error)) || error instanceof TransactionRejectedError
-
-      if (!userRejected) {
-        setSwapState({
-          attemptingTxn: false,
-          tradeToConfirm,
-          swapErrorMessage: typeof error === 'string' ? error : error?.message,
-          txHash: undefined,
-        })
-      }
-
-      throw error
-    }
-  }, [swapCallback, tradeToConfirm])
-
   const handleAcceptChanges = useCallback(() => {
-    setSwapState({ tradeToConfirm: rfqTrade.trade, swapErrorMessage, txHash, attemptingTxn })
-  }, [attemptingTxn, swapErrorMessage, rfqTrade, txHash, setSwapState])
+    setTradeToConfirm(rfqTrade.trade)
+    resetState()
+  }, [rfqTrade.trade, resetState])
 
   const handleConfirmDismiss = useCallback(() => {
-    setSwapState({ tradeToConfirm, attemptingTxn, swapErrorMessage, txHash })
+    resetState()
     // if there was a tx hash, we want to clear the input
     if (txHash) {
       onUserInput(Field.INPUT, '')
     }
-  }, [attemptingTxn, onUserInput, swapErrorMessage, tradeToConfirm, txHash, setSwapState])
+  }, [resetState, txHash, onUserInput])
 
   // Modals
   const [indirectlyOpenConfirmModalState, setIndirectlyOpenConfirmModalState] = useState(false)
@@ -140,35 +109,18 @@ SwapCommitButtonPropsType) {
     />,
   )
 
-  const amountToApprove = useMemo(() => {
-    return mmTradeInfo?.slippageAdjustedAmounts[Field.INPUT]
-  }, [mmTradeInfo])
-  const mmSpender = useMemo(() => {
-    return mmTradeInfo?.routerAddress as Address | undefined
-  }, [mmTradeInfo])
-  const { execute: onStep, approvalState } = usePermitOrApprove(amountToApprove, mmSpender, onConfirmSwap)
-
-  const { confirmModalState, pendingModalSteps, resetConfirmModalState, startSwap } = useConfirmModalState(
-    onStep,
-    amountToApprove,
-    approvalState,
-    undefined,
-    undefined,
-    mmSpender,
-  )
-
   const [onPresentConfirmModal] = useModal(
     <ConfirmSwapModal
       isMM
       trade={rfqTrade.trade} // show the info while refresh RFQ
       originalTrade={tradeToConfirm}
       txHash={txHash}
-      confirmModalState={confirmModalState}
-      pendingModalSteps={pendingModalSteps}
-      onConfirm={startSwap}
+      confirmModalState={confirmState}
+      pendingModalSteps={confirmSteps ?? []}
+      onConfirm={callToAction}
       currencyBalances={currencyBalances}
       isRFQReady={Boolean(rfqTrade.rfq) && !rfqTrade.isLoading}
-      swapErrorMessage={swapErrorMessage || (!rfqTrade.trade ? t('Unable request a quote') : undefined)}
+      swapErrorMessage={errorMessage || (!rfqTrade.trade ? t('Unable request a quote') : undefined)}
       onAcceptChanges={handleAcceptChanges}
       customOnDismiss={handleConfirmDismiss}
       openSettingModal={onPresentSettingsModal}
@@ -180,32 +132,24 @@ SwapCommitButtonPropsType) {
   // End Modals
 
   const onSwapHandler = useCallback(() => {
-    setSwapState({
-      tradeToConfirm: rfqTrade.trade,
-      attemptingTxn: false,
-      swapErrorMessage: undefined,
-      txHash: undefined,
-    })
-    resetConfirmModalState()
+    setTradeToConfirm(rfqTrade.trade)
+    resetState()
 
     if (isExpertMode) {
-      startSwap()
+      callToAction()
     }
     onPresentConfirmModal()
     logGTMClickSwapEvent()
-  }, [rfqTrade.trade, resetConfirmModalState, isExpertMode, onPresentConfirmModal, startSwap])
+  }, [rfqTrade.trade, resetState, isExpertMode, onPresentConfirmModal, callToAction])
 
   // useEffect
   useEffect(() => {
     if (indirectlyOpenConfirmModalState) {
       setIndirectlyOpenConfirmModalState(false)
-      setSwapState((state) => ({
-        ...state,
-        swapErrorMessage: undefined,
-      }))
+      resetState()
       onPresentConfirmModal()
     }
-  }, [indirectlyOpenConfirmModalState, onPresentConfirmModal, setSwapState])
+  }, [indirectlyOpenConfirmModalState, onPresentConfirmModal, resetState])
 
   if (swapIsUnsupported) {
     return (
@@ -234,7 +178,7 @@ SwapCommitButtonPropsType) {
       width="100%"
       id="swap-button"
       variant="primary"
-      disabled={!rfqTrade.rfq || !isValid || !!swapCallbackError}
+      disabled={!rfqTrade.rfq || !isValid || !!errorMessage}
       onClick={onSwapHandler}
     >
       {parseMMError(swapInputError) || t('Swap')}
