+++ apps/web/src/views/Swap/MMLinkPools/hooks/useMMConfirmModalState.ts
@@ -7,7 +7,7 @@ import { useCallback, useEffect, useMemo, useRef, useState } from 'react'
 import { publicClient } from 'utils/client'
 import { UserUnexpectedTxError } from 'utils/errors'
 import { Address, Hex } from 'viem'
-import { ConfirmAction } from 'views/Swap/V3Swap/hooks/useConfirmModalStateV2'
+import { ConfirmAction } from 'views/Swap/V3Swap/hooks/useConfirmModalState'
 import { userRejectedError } from 'views/Swap/V3Swap/hooks/useSendSwapTransaction'
 import { useApprove, useApproveRequires } from './useApprove'
 import { MMSwapCall } from './useSwapCallArguments'
