+++ apps/web/src/views/Swap/V3Swap/containers/MMCommitButton.tsx
@@ -7,7 +7,7 @@ import { useMemo } from 'react'
 import { Field } from 'state/swap/actions'
 import { useSwapState } from 'state/swap/hooks'
 import { useSwapActionHandlers } from 'state/swap/useSwapActionHandlers'
-import { MMSwapCommitButton } from 'views/Swap/MMLinkPools/components/MMCommitButtonV2'
+import { MMSwapCommitButton } from 'views/Swap/MMLinkPools/components/MMCommitButton'
 import { useAccount } from 'wagmi'
 
 export function MMCommitButton({ mmOrderBookTrade, mmRFQTrade, mmQuoteExpiryRemainingSec, mmTradeInfo }) {
