--- apps/web/src/views/Swap/V3Swap/containers/SwapCommitButtonV2.tsx
@@ -1,286 +1,0 @@
-import { TradeType } from '@pancakeswap/sdk'
-import { SmartRouterTrade } from '@pancakeswap/smart-router/evm'
-import { Currency, CurrencyAmount, Token } from '@pancakeswap/swap-sdk-core'
-import { AutoColumn, Box, Button, Dots, Message, MessageText, Text, useModal } from '@pancakeswap/uikit'
-import React, { memo, useCallback, useEffect, useMemo, useState } from 'react'
-
-import { useTranslation } from '@pancakeswap/localization'
-import { getUniversalRouterAddress } from '@pancakeswap/universal-router-sdk'
-import { ConfirmModalState } from '@pancakeswap/widgets-internal'
-import { GreyCard } from 'components/Card'
-import { CommitButton } from 'components/CommitButton'
-import ConnectWalletButton from 'components/ConnectWalletButton'
-import { AutoRow } from 'components/Layout/Row'
-import SettingsModal, { RoutingSettingsButton, withCustomOnDismiss } from 'components/Menu/GlobalSettings/SettingsModal'
-import { SettingsMode } from 'components/Menu/GlobalSettings/types'
-import { BIG_INT_ZERO } from 'config/constants/exchange'
-import { useCurrency } from 'hooks/Tokens'
-import { useIsTransactionUnsupported } from 'hooks/Trades'
-import useWrapCallback, { WrapType } from 'hooks/useWrapCallback'
-import { Field } from 'state/swap/actions'
-import { useSwapState } from 'state/swap/hooks'
-import { useSwapActionHandlers } from 'state/swap/useSwapActionHandlers'
-import { useRoutingSettingChanged } from 'state/user/smartRouter'
-import { useCurrencyBalances } from 'state/wallet/hooks'
-import { logGTMClickSwapEvent } from 'utils/customGTMEventTracking'
-import { warningSeverity } from 'utils/exchange'
-import { useAccount, useChainId } from 'wagmi'
-import { useParsedAmounts, useSlippageAdjustedAmounts, useSwapInputError } from '../hooks'
-import { useConfirmModalStateV2 } from '../hooks/useConfirmModalStateV2'
-import { useSwapConfig } from '../hooks/useSwapConfig'
-import { useSwapCurrency } from '../hooks/useSwapCurrency'
-import { computeTradePriceBreakdown } from '../utils/exchange'
-import { ConfirmSwapModal } from './ConfirmSwapModalV2'
-
-const SettingsModalWithCustomDismiss = withCustomOnDismiss(SettingsModal)
-
-interface SwapCommitButtonPropsType {
-  trade?: SmartRouterTrade<TradeType>
-  tradeError?: Error
-  tradeLoading?: boolean
-}
-
-const useSettingModal = (onDismiss) => {
-  const [openSettingsModal] = useModal(
-    <SettingsModalWithCustomDismiss customOnDismiss={onDismiss} mode={SettingsMode.SWAP_LIQUIDITY} />,
-  )
-  return openSettingsModal
-}
-
-const useSwapCurrencies = () => {
-  const {
-    [Field.INPUT]: { currencyId: inputCurrencyId },
-    [Field.OUTPUT]: { currencyId: outputCurrencyId },
-  } = useSwapState()
-  const inputCurrency = useCurrency(inputCurrencyId) as Currency
-  const outputCurrency = useCurrency(outputCurrencyId) as Currency
-  return { inputCurrency, outputCurrency }
-}
-
-const WrapCommitButtonReplace: React.FC<React.PropsWithChildren> = ({ children }) => {
-  const { t } = useTranslation()
-  const { inputCurrency, outputCurrency } = useSwapCurrencies()
-  const { typedValue } = useSwapState()
-  const {
-    wrapType,
-    execute: onWrap,
-    inputError: wrapInputError,
-  } = useWrapCallback(inputCurrency, outputCurrency, typedValue)
-  const showWrap = wrapType !== WrapType.NOT_APPLICABLE
-
-  if (!showWrap) return children
-
-  return (
-    <CommitButton width="100%" disabled={Boolean(wrapInputError)} onClick={onWrap}>
-      {wrapInputError ?? (wrapType === WrapType.WRAP ? t('Wrap') : wrapType === WrapType.UNWRAP ? t('Unwrap') : null)}
-    </CommitButton>
-  )
-}
-
-const ConnectButtonReplace = ({ children }) => {
-  const { address: account } = useAccount()
-
-  if (!account) {
-    return <ConnectWalletButton width="100%" />
-  }
-  return children
-}
-
-const UnsupportedSwapButtonReplace = ({ children }) => {
-  const { t } = useTranslation()
-  const { inputCurrency, outputCurrency } = useSwapCurrencies()
-  const swapIsUnsupported = useIsTransactionUnsupported(inputCurrency, outputCurrency)
-
-  if (swapIsUnsupported) {
-    return (
-      <Button width="100%" disabled>
-        {t('Unsupported Asset')}
-      </Button>
-    )
-  }
-  return children
-}
-
-export const SwapCommitButton: React.FC<SwapCommitButtonPropsType> = (props) => {
-  return (
-    <UnsupportedSwapButtonReplace>
-      <ConnectButtonReplace>
-        <WrapCommitButtonReplace>
-          <SwapCommitButtonInner {...props} />
-        </WrapCommitButtonReplace>
-      </ConnectButtonReplace>
-    </UnsupportedSwapButtonReplace>
-  )
-}
-
-const SwapCommitButtonInner = memo(function SwapCommitButtonInner({
-  trade,
-  tradeError,
-  tradeLoading,
-}: SwapCommitButtonPropsType) {
-  const { address: account } = useAccount()
-  const { t } = useTranslation()
-  const chainId = useChainId()
-  // form data
-  const { independentField } = useSwapState()
-  const [inputCurrency, outputCurrency] = useSwapCurrency()
-  const { isExpertMode } = useSwapConfig()
-
-  const slippageAdjustedAmounts = useSlippageAdjustedAmounts(trade)
-  const amountToApprove = useMemo(
-    () => (inputCurrency?.isNative ? undefined : slippageAdjustedAmounts[Field.INPUT]),
-    [inputCurrency?.isNative, slippageAdjustedAmounts],
-  )
-
-  const tradePriceBreakdown = useMemo(() => computeTradePriceBreakdown(trade), [trade])
-  // warnings on slippage
-  const priceImpactSeverity = warningSeverity(
-    tradePriceBreakdown ? tradePriceBreakdown.priceImpactWithoutFee : undefined,
-  )
-
-  const relevantTokenBalances = useCurrencyBalances(account ?? undefined, [
-    inputCurrency ?? undefined,
-    outputCurrency ?? undefined,
-  ])
-  const currencyBalances = {
-    [Field.INPUT]: relevantTokenBalances[0],
-    [Field.OUTPUT]: relevantTokenBalances[1],
-  }
-  const parsedAmounts = useParsedAmounts(trade, currencyBalances, false)
-  const parsedIndependentFieldAmount = parsedAmounts[independentField]
-  const swapInputError = useSwapInputError(trade, currencyBalances)
-  const [tradeToConfirm, setTradeToConfirm] = useState<SmartRouterTrade<TradeType> | undefined>(undefined)
-  const [indirectlyOpenConfirmModalState, setIndirectlyOpenConfirmModalState] = useState(false)
-
-  const { callToAction, confirmState, txHash, confirmSteps, errorMessage, resetState } = useConfirmModalStateV2(
-    trade,
-    inputCurrency?.isToken ? (amountToApprove as CurrencyAmount<Token>) : undefined,
-    getUniversalRouterAddress(chainId),
-  )
-
-  const { onUserInput } = useSwapActionHandlers()
-  const reset = useCallback(() => {
-    if (confirmState === ConfirmModalState.COMPLETED) {
-      onUserInput(Field.INPUT, '')
-    }
-    resetState()
-  }, [confirmState, onUserInput, resetState])
-
-  const handleAcceptChanges = useCallback(() => {
-    setTradeToConfirm(trade)
-  }, [trade])
-
-  const noRoute = useMemo(() => !((trade?.routes?.length ?? 0) > 0) || tradeError, [trade?.routes?.length, tradeError])
-  const isValid = useMemo(() => !swapInputError && !tradeLoading, [swapInputError, tradeLoading])
-  const disabled = useMemo(
-    () => !isValid || (priceImpactSeverity > 3 && !isExpertMode),
-    [isExpertMode, isValid, priceImpactSeverity],
-  )
-
-  const userHasSpecifiedInputOutput = Boolean(
-    inputCurrency && outputCurrency && parsedIndependentFieldAmount?.greaterThan(BIG_INT_ZERO),
-  )
-
-  // modals
-  const onSettingModalDismiss = useCallback(() => {
-    setIndirectlyOpenConfirmModalState(true)
-  }, [])
-  const openSettingModal = useSettingModal(onSettingModalDismiss)
-  const [openConfirmSwapModal] = useModal(
-    <ConfirmSwapModal
-      trade={trade}
-      originalTrade={tradeToConfirm}
-      txHash={txHash}
-      confirmModalState={confirmState}
-      pendingModalSteps={confirmSteps ?? []}
-      swapErrorMessage={errorMessage}
-      currencyBalances={currencyBalances}
-      onAcceptChanges={handleAcceptChanges}
-      onConfirm={callToAction}
-      openSettingModal={openSettingModal}
-      customOnDismiss={reset}
-    />,
-    true,
-    true,
-    'confirmSwapModal',
-  )
-
-  const handleSwap = useCallback(() => {
-    setTradeToConfirm(trade)
-    resetState()
-
-    // if expert mode turn-on, will not show preview modal
-    // start swap directly
-    if (isExpertMode) {
-      callToAction()
-    }
-
-    openConfirmSwapModal()
-    logGTMClickSwapEvent()
-  }, [callToAction, isExpertMode, openConfirmSwapModal, resetState, trade])
-
-  useEffect(() => {
-    if (indirectlyOpenConfirmModalState) {
-      setIndirectlyOpenConfirmModalState(false)
-      openConfirmSwapModal()
-    }
-  }, [indirectlyOpenConfirmModalState, openConfirmSwapModal])
-
-  if (noRoute && userHasSpecifiedInputOutput && !tradeLoading) {
-    return <ResetRoutesButton />
-  }
-
-  return (
-    <Box mt="0.25rem">
-      <CommitButton
-        id="swap-button"
-        width="100%"
-        variant={isValid && priceImpactSeverity > 2 && !errorMessage ? 'danger' : 'primary'}
-        disabled={disabled}
-        onClick={handleSwap}
-      >
-        {swapInputError ||
-          (tradeLoading && <Dots>{t('Searching For The Best Price')}</Dots>) ||
-          (priceImpactSeverity > 3 && !isExpertMode
-            ? t('Price Impact Too High')
-            : priceImpactSeverity > 2
-            ? t('Swap Anyway')
-            : t('Swap'))}
-      </CommitButton>
-    </Box>
-  )
-})
-
-const ResetRoutesButton = () => {
-  const { t } = useTranslation()
-  const [isRoutingSettingChange, resetRoutingSetting] = useRoutingSettingChanged()
-  return (
-    <AutoColumn gap="12px">
-      <GreyCard style={{ textAlign: 'center', padding: '0.75rem' }}>
-        <Text color="textSubtle">{t('Insufficient liquidity for this trade.')}</Text>
-      </GreyCard>
-      {isRoutingSettingChange && (
-        <Message variant="warning" icon={<></>}>
-          <AutoColumn gap="8px">
-            <MessageText>{t('Unable to establish trading route due to customized routing.')}</MessageText>
-            <AutoRow gap="4px">
-              <RoutingSettingsButton
-                buttonProps={{
-                  scale: 'xs',
-                  p: 0,
-                }}
-                showRedDot={false}
-              >
-                {t('Check your settings')}
-              </RoutingSettingsButton>
-              <MessageText>{t('or')}</MessageText>
-              <Button variant="text" scale="xs" p="0" onClick={resetRoutingSetting}>
-                {t('Reset to default')}
-              </Button>
-            </AutoRow>
-          </AutoColumn>
-        </Message>
-      )}
-    </AutoColumn>
-  )
-}
